requirements.txt
accelerate==0.16.0
albumentations==1.3.0
diffusers==0.11.1
gradio==3.20.1
huggingface-hub==0.13.0
opencv-python-headless==4.7.0.68
Pillow==9.4.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.65.0
transformers==4.26.1